为避免修改卡密导致数据库崩溃做出系列调整
1、禁用CardEditor修改卡密的功能，以后只能在创建卡片的时候确定卡密，创建好的卡片不能修改卡密。
2、修改了DialogService类库，让其支持直接根据ViewModel类型打开窗体并返回对应的ViewModel。

+++ YGOProDevelop/Service/ICDBService.cs
@@ -6,7 +6,7 @@ namespace YGOProDevelop.Service
     public interface ICDBService
     {
         void Add(YGOProDevelop.Model.datas card);
-        bool IsIDExisted(int id);
+        bool IsIDExisted(long id);
         void Open(string filePath);
         void Remove(YGOProDevelop.Model.datas card);
         void Replace(YGOProDevelop.Model.datas from, YGOProDevelop.Model.datas to);
